+++ public/index.html
@@ -24,6 +24,14 @@
   <!-- Сортировка -->
 
   <!-- Контент -->
+  <section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+
+      <div class="films-list__container">
+      </div>  
+    </section>
+  </section>
 </main>
