+++ src/view/show-more-button-view.js
@@ -1,0 +1,4 @@
+export const createShowMoreButton = () => (
+  `<button class="films-list__show-more">Show more
+  </button>`
+);
